[#3]/refactor TutorialContents 중복 제거, tutorial 폴더 생성
